refactor methods to utility folder
